Protect against NPE if the DB wasn't initialized during HModel init

+++ org.hawk.osgiserver/src/org/hawk/osgiserver/HModel.java
@@ -101,7 +101,9 @@ public class HModel {
 			console.printerrln(e);
 
 			try {
-				db.delete();
+				if (db != null) {
+					db.delete();
+				}
 			} catch (Exception e2) {
 				throw e2;
 			}
